+++ src/js/index.js
@@ -1,3 +1,1 @@
 import css from '../css/main.css';
-// import './recipes'
-// import './ingredient'
